Bug#54242 ndb native default values break ndb_restore
Ndb-native support for default values allows column definitions in the Ndb 
data nodes to include default value information.

Currently that information must be constant.

In MySQL, the TIMESTAMP type can have a 'default value' which is automatically
set to the CURRENT_TIMESTAMP at INSERT or on row UPDATE time.  This cannot 
currently be supported natively by the Ndb storage engine.
Currently, the Ndb storage engine treats TIMESTAMP columns with DEFAULT CURRENT_TIMESTAMP
as if they were DEFAULT 0.

This patch returns the TIMESTAMP behaviour to be the same as it was before
the native default support was added, except in the case where a constant
TIMESTAMP default (e.g. 0 or 20100607174832) is set.

The ndb_native_default_support testcase is enhanced to record the expected
behaviour.