BUG#53060: LooseScan semijoin strategy does not return all rows
Problem: In sub_select(), loosescan_match_tab->found_match is not reset when
moving to the next key value. This means that a key value for which a match 
has not yet been found, may be copied into loosescan_buf.  All later records
with the same key value will then be skipped.

Solution: Reset loosescan_match_tab->found_match to FALSE when the current
key does not match the previous key.  Now a new value is not copied and used
for comparison until a new match has been found for the duplicate-generating
range.

Additional change:  Moved the whole LooseScan logic to evaluate_join_record
to colocate it with similar duplicate-avoiding logic for other semijoin
strategies.