BUG#46554: RBR: pending rows event can be flushed too soon sometimes.
While packing a record into an rows event, the server performs
several checks to see if a new rows event is needed. Among the
conditions checked, there's one that deals with bitmaps from the
pending event ('p') and the table ('t') that the server accessed
for the current record. If they are different, then 'p' is
flushed and a new pending event, 'np', is created. If not, then
'p' is used. However, the only bitmaps compared are
't'->write_set against 'p'->get_cols(), which in some cases,
means we are comparing bitmaps related to write_sets (the ones
from 't') against bitmaps related to read_sets (the ones from
'p')!  In current versions of the server this issue is masked (as
both read and write sets are always fully set).

On the other hand, for 6.0 codebase this was not the case
because if the table has a PK, only some columns will be flagged
in the read_set. In such cases the write_set and the read_set
will most likely not match causing the pending to be flushed.
Additionally, the write_set is meaningless for delete events and
the read_set is meaningless for write events! So comparison must
take into account the type of the event to choose what bitmap to
take from the table to make the comparison.

As stated, this was the case for 6.0 codebase, but also for trees
patched with WL#5092, because the user is given the option
to choose whether to log partial (not all columns will be flagged
in the read and write set) or full rows. If the user chooses to
log partial rows, the problem may surface.

Finally, This seems to be in fact a regression introduced by
patch for BUG#33055 plus a post-merge fix that is related:
 - sp1r-mkindahl@dl145h.mysql.com-20080211101647-37854

Since BUG#33055 was backported to WL#5092 tree, this regression
tagged along.

To fix this problem, we deploy a member function in
Rows_log_event that is used to perform the correct comparison of
the pending and table bitmaps.