BUG#53103 - MTR test ps crashes in optimize_cond() when             running with --debug
The problem was that when subselect_hash_sj_engine was used to 
materialize a subselect, the subselect was not prepared for 
each execution of a PS. Because of this, Items in the subselect
were not fixed on the second execution of the PS.

The fix is to call materialize_engine->prepare() in 
subselect_hash_sj_engine::prepare() method. From comments 
in the class, this seems to have been the intention all along.