Post-push fix for WL#2775.
The type of the repository when it is a table may be changed at anytime.
For that reason the function that checks its transactional properties
cannot use a constant defined in the start up. 

The code now opens the table and checks its type, i.e. transactional or
non-transactional, whenever is necessary.