Post-push fix for WL#2775.
Improved the Rpl_info_table_access::find_info, Rpl_info_table_access::open_table and
Rpl_info_factory::decide_repository. Along with that the the content of the master
info and relay log info are now always flushed when the relay log is rotated, the
CHANGE MASTER is executed or the SQL-I/O Thread is stopped.

In particular, for the relay log info, its contents are flushed after processing any
statement written to the binary log without a transaction context and after processing
commit or rollback, i.e. a Query_log_event.