Bug#52305 : Customizing storage of server id inside the anyValue
Add server-id-bits option to MySQLD

A general server-id-bits option is added to MySQLD initialisation options.
This option indicates the number of lsbs within the 32-bit ServerId which actually contain the server-id.
Indicating that the actual server-id uses < 32 bits allows some of the remaining bits to be used for other purposes.  When extracting the actual server-id for purposes such as replication-loop-detection, the server will ignore the irrelevant bits.
If the configured server-id cannot be expressed in server-id-bits then the MySQLD will refuse to start.
The server-id bits option is used mask out the irrelevant bits of the serverid in the IO and SQL threads when deciding whether an event should be ignored based on serverid.
The event's serverid is used to set the slave thread's thd->server_id in the normal way, and a new thd variable, unmasked_server_id carries the full 32-bits of the received server_id.
In cases where server_id_bits is < 32, handlers can look at thd->unmasked_server_id to get the non-server-id data associated with the event.

The mysqlbinlog tool gets a new option --server-id-bits which controls how it interprets the server ids in the Binlogs it reads.  As with MySQLD this defaults to 32.  Where the Binlog was written by a MySQLD with server-id-bits < 32 and some useful info in the msbs, running mysqlbinlog with server-id-bits of 32 allows this data to be seen.