Bug#52305 Customizing storage of server id inside the anyValue
Test UserData part of AnyValue

Add a DBUG-only env variable to trigger MySQLD to set the user portion of AnyValue to all 1s.
Modify my.cnf for some AnyValue-sensitive testcases to set a 24-bit user portion which will be set to all 1s.
This gives coverage of the case(s) where the User portion is non-zero, checking that the system still functions as expected w.r.t. ServerId, NoLogging etc.

The existing repl_ndb_slave_lsu testcase is modified to also use the mysqlbinlog tool to look at the events recorded.
A new variant of it called rpl_ndb_slave_lsu_anyval runs with the top 24 bits set to 1 and server-id-bits set to 7.  The testcase ensures that :
 - Server-ids are correctly extracted from the composite serverid
 - Composite serverids are binlogged
 - Composite serverids pass through the slave and can be binlogged
 - mysqlbinlog can be used to see composite server id values.