BUG#54866: Partially failed REVOKE not binlogged, causes inconsistency or replication abort
The server was skipping binary logging for some half complete
GRANT and REVOKE operations.

We fix this by writing to the binary log GRANT/REVOKE operations,
those that were successful (nothing new here) and those that may
have changed master state, even if they ended up in
error. Additionally, for partially succeeded GRANT/REVOKE
statements we log the error code the master got from the
execution. This way we give the slave a chance to replay the
statement, get the same error, and the same side effects from the
unsuccessful execution as the master did.

Finally, for failed REVOKE/GRANT statements that don't change 
state, we don't log them.