Bug#49453: "re-execution of prepared statement with             view and semijoin crashes"
When semijoin was applied on a view, the replaced items were
pulled out instead of Item_field since the items referring
to view columns had already been fixed. However, the semijoin
nest still stored the item list in sj_subq_pred, and on next
execution of the PS, this item list was reset during parsing.
Since this list is not attached to the query, Item_fields 
referring to view columns were not replaced with 
Item_direct_view_ref as would happen in normal execution.

The fix for this is to use the sj_inner_exprs list of the 
semijoin nest instead of the item list. The sj_inner_exprs list
contains the resolved items in the subquery.