Bug#50489: another segfault in fix_semijoin_strategies...
This particular problem occurs when we try to execute a query multiple
times with different optimizer switch settings. On first execution,
semijoin materialization strategy is a candidate and an
SJ_MATERIALIZATION_INFO object is created. On second execution,
materialization strategy is not a candidate, but there is a dangling
pointer to the SJ_MATERIALIZATION_INFO object created in the prior
execution. The pointer is followed, and we suffer a segfault.

The solution chosen here is to refactor the SJ_MATERIALIZATION_INFO
class so that optimization information is put into a new struct
Semijoin_mat_optimize which is embedded in struct st_nested_join.
SJ_MATERIALIZATION_INFO itself is renamed to Semijoin_mat_exec.

If semijoin materialization is an allowed strategy and materialization
is possible for the semijoin in question, the Semijoin_mat_optimize object
is filled in and an st_position object is created and filled with
information about the join strategy for the materialization.

Later in the optimization, the materialization strategy is compared
with other semijoin strategies, and, if selected, an
Semijoin_mat_exec object is created and attached to the join nest
representing the semijoin operation.
The object is then filled with data relevant for semijoin execution,
such as number of tables involved, whether this is a scan, information
about temporary table for materialization, etc.

NOTE. This bug no longer occurs on the current souce code base.
      It does however fix bug#46744, which is re-introduced by the
      fix for bug#43768, and it should also be useful because of the
      refactoring effect on the semijoin materialization strategy.

mysql-test/r/optimizer_switch.result
  Added results for test case for bug#50489

mysql-test/t/optimizer_switch.test
  Added test case for bug#50489

sql/sql_class.h
  Optimization data are removed from class SJ_MATERIALIZATION_INFO.
  SJ_MATERIALIZATION_INFO is renamed to Semijoin_mat_exec.

sql/sql_select.cc
  optimize_semijoin_nests() will no longer create
  SJ_MATERIALIZATION_INFO objects. Instead it populates the
  Semijoin_mat_optimize struct inside the nested_join object with estimated
  cost for the materialized semijoin.

  fix_semijoin_strategies_for_picked_join_order() is given a boolean
  return value. The return value is necessary because the function may
  create Semijoin_mat_exec objects, hence it must be able to
  return error conditions. Notice that these objects are created and
  populated only if a semijoin materialization strategy is actually
  chosen. Reset of sjm.positions is performed in this function.

  advance_sj_state() has been slightly simplified: the "emb_sj_nest"
  is now assigned when the function is entered, in addition to changes
  that was necessitated by the splitting up of SJ_MATERIALIZATION_INFO.

  Reset of the pointer sj_nest->sj_mat_exec is added to JOIN::destroy().

sql/sql_select.h
  List with name sjm_info_list is renamed to sjm_exec_list.

sql/sql_test.cc
  New interface for function print_sjm(), made necessary by the splitting
  of SJ_MATERIALIZATION_INFO.

sql/sql_test.h
  New interface for function print_sjm().

sql/table.h
  Defines struct Semijoin_mat_optimize containing estimated cost for a
  materialized semijoin operation.
  The struct is added as a member in struct st_nested_join.