Bug#55121 error 839 'Illegal null attribute' from NDB for fields with default value
User supplying an (ignored) default on the Blob column causes the
NO_DEFAULT_VALUE_FLAG *not* to be set on the column, so we incorrectly 
assume that it must have a native default and send no data for it.

Ndb handler is modified to never assume that a Blob (or Geom etc.) has 
a native default value.