Bug #56006
there were failure with rpl_checksum_cache in MBL::write_cache() caused by
an opeque issue with int4store expansion on 64 bit platforms (linux in particular).

The core reason of int4store issue is yet unknown (a bug to be reported). However, there is a simple workaround to not invoke the macro inside another macro which is implemented to satisfy the test