Bug #53167  	Seconds_Behind_Master is wrong after start slave
When a slave server has an existing relay log and the user issues
START SLAVE, the output from "SHOW SLAVE STATUS" shows a wrong
value for Seconds_Behind_Master until the first event from the relay
log has finished executing. The displayed value for Seconds_Behind_Master
is calculated based on the timestamp of the Format_Description_Log_Event,
which is reinserted into stream of events in RL when RL rotates. This can
be a very big number if the previous relay log was created long ago. And
the first FD generated by slave, which refers to incomparable time generated
on slave.

Do not assign the Timestamp of the Format_Description_Log_Event to
last_master_timestamp when executing the Format_Description_Log_Event
on slave. Then the peak value will disappear.