Bug#48213 Materialized subselect crashes if using GEOMETRY type
The problem occurred because during execution of MaterializeLookup
strategy for semi-join, a materialized table was created which
contained a GEOMETRY column, a specialized BLOB column.  This caused
an segmentation fault because such tables will have extra columns to record
NULL information for the BLOB column, and the semi-join code was not 
prepared for that.

The solution is to disable MaterializeLookup strategy when
Blob/Geometry columns would need to be materialized.  Blob columns
cannot be used for index look-up anyway, so it does not makes sense to
use MaterializeLookup.

This fix also implies that it is detected earlier that subquery
materialization can not be used either.  This means that code to
revert the decision to use subquery materialization is no longer
needed, and this has been removed.