BUG#55704 incorrect binlogging of savepoints break slaves!
The replication was ignoring that a savepoint was valid within a transaction unless
it is released, implicit or explicitly, and was removing its occurrence from the
binary log when a 'rollback to' was issued.

To fix the problem, we always keep the savepoint defintion in the binary log when
rolling back to the savepoint.