BUG#56883: rpl_row_ignorable_event fails on valgrind run
The buffer used in mysql_client_binlog_statement is never
initialized and may contain several events data in it. If an
event is processed and its payload is handled as a null
terminated string, then conditional jumps depending on
uninitialized values may ocur. Such cases happen if there is 
access outside this event portion of the buffer used for 
storing its data. This was the case for Rows_query_log_event.

We fix this by replacing my_snprintf with strmake (which copies
only the specified len bytes from str and sets 
*(my_rows_query+len)= '\0'.