Bug #56492  	a flaw in String::chop()
String::chop() used postfix decrement inside a subscript operator

  Ptr[str_length--]= '\0'; 

with intent to subtract 1 from the length and put \0 into the indexed by the *new* value
of `str_length' position.
While the first part succeeds - the length is changed - the 2nd does not and \0 remains
(if it was set there before) at the same position (old `str_length') after the method as
it was before the invocation.

Logics of chop() corrected and simplified, an assert is added to
confirm the correct execution ever.