Bug #56913    rpl_row_ignorable_event fails sporadically on next-mr-bugfixing codebase
In RBR, The rows are inserted to a queue by the thread executing
the 'INSERT DELAYED' statement, and are taken out from the queue
by the handler thread to do the real insertion. Because these two
threads are running in parallel, there is a possibility that they
are run in an interleaved manner, and result in different number
of table_map and rows events.

Added a debug option for the test to make the binlog of multi
'INSERT DELAYED ...' stmt stable by forcing every value is
executed into one execution series, and then each value will
be binlogged into a separate rows event with its table map
event. And the original SQL statement will be displayed just
for the first separate rows event because the query is NULL
in other separate rows.