This commit fixes two errors related to pushed queries involving varchar types:
1. Some queries would fail with the error message: 'Character Parameter was
right truncated'. This was due to an attempt to convert string from a format
with 2-byte length field to 1-byte length field when the string already had
a one-byte length field.

2. Varchar values that were parameters to non-root pushed operations would
not get the one- or two-byte length fields they should have. The SPJ block would
then either crash (on an ndbrequire) or misinterpret the string.