Bug#57708: EXCHANGE PARTITION misuses IGNORE
First problem was that IGNORE was used for skipping the whole
validation step (check of all rows in the table to fit
into the partition) instead of changing every row error into a
warning.

Second problem was that IGNORE opened up the possibility for
bugs like 55944.

Solution was to remove the IGNORE option from EXCHANGE PARTITION
WITH TABLE (WL#4445).