BUG#53386: HA_ERR_CORRUPT_EVENT or debug assertion with different table on slave
When the number of raw bytes used for a CHAR field exceeds 255
two bytes are actually used while packing the length of the
field. CHAR fields can exceed the raw 255 byte length, for
instance, if using multi-byte character sets.

The problem is that when calculating the raw length of the field,
at unpacking time, only one byte was assumed to be used
always. Therefore, the number of bytes used would not add up,
triggering the assertion.

We fix this, by making the member function that calculates the
raw byte length for a given field to take into account the actual
number of bytes used while packing the field length.